an example of passing by reference
We create a function that will be invoked with our pass_by_reference value. Within said function we display the value of the parameter. The parameter is preceded by an ampersand during definition. In our main function we pass a variable we create to the invocation of the pass_by_ref function resulting in the value being changed from its original value. We log the results of param and var1 to the console. (var1: 1 param: 1 param: 2 var1: 2)